Move directory functions to DanceSaver
Before, several functions to create directories for DanceSaver were in
the main dance.py script, but it seems to make more sense if they are in
the DanceSaver library.
